Snaplytics/requirements.txt
asgiref==3.11.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==1.26.4
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plumbum==1.10.0
ply==3.11
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.7.2
scikit-surprise==1.1.4
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
surprise==0.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
xgboost==3.2.0
